RISC-V: Implement riscv_minimal_hwprobe_feature_bits
This patch implements the riscv_minimal_hwprobe_feature_bits feature
for the RISC-V target. The feature bits are defined in the previous
patch [1] to provide bitmasks of ISA extensions that defined in RISC-V
C-API. Thus, we need a function to generate the feature bits for IFUNC
resolver to dispatch between different functions based on the hardware
features. The final version of the target_clones support on RISC-V is
still under development, I am working on it.

The minimal feature bits means to use the earliest extension appeard in
the Linux hwprobe to cover the given ISA string. To allow older kernels
without some implied extensions probe to run the FMV dispatcher
correctly.

For example, V implies Zve32x, but Zve32x appears in the Linux kernel
since v6.11. If we use isa string directly to generate FMV dispatcher
with functions with "arch=+v" extension, since we have V implied the
Zve32x, FMV dispatcher will check if the Zve32x extension is supported
by the host. If the Linux kernel is older than v6.11, the FMV dispatcher
will fail to detect the Zve32x extension even it already implies by the
V extension, thus making the FMV dispatcher fail to dispatch the correct
function.

Thus, we need to generate the minimal feature bits to cover the given
ISA string to allow the FMV dispatcher to work correctly on older
kernels.

[1] https://patchwork.sourceware.org/project/gcc/patch/20241003182256.1765569-1-chenyangyu@isrc.iscas.ac.cn/

gcc/ChangeLog:

	* common/config/riscv/riscv-common.cc
	(struct riscv_ext_bitmask_table_t): New struct.
	(riscv_minimal_hwprobe_feature_bits): New function.
	* config/riscv/riscv-subset.h (GCC_RISCV_SUBSET_H):
	(riscv_minimal_hwprobe_feature_bits): Declare the function.
	* common/config/riscv/feature_bits.h: New file.