Redirect to specific target based on TARGET_VERSION_COMPATIBLE
This commit introduces a new target hook, TARGET_VERSION_COMPATIBLE,
enabling the redirection of specific target clones based on version
compatibility. Additionally, it modifies the existing redirection
mechanism to support this new hook, which identifies the most
prioritized MV clone compatible with the current caller version and
redirects to it.

This mechanism has been discussed in a previous patch [1].

[1] https://patchwork.sourceware.org/comment/197172/

Signed-off-by: Yangyu Chen <cyy@cyyself.name>

gcc/ChangeLog:

	* doc/tm.texi: New Target Hook TARGET_VERSION_COMPATIBLE.
	* doc/tm.texi.in: Ditto.
	* target.def: Ditto.
	* multiple_target.cc (redirect_to_specific_clone): Support
	redirect_to_specific_clone with TARGET_VERSION_COMPATIBLE.
	(ipa_target_clone): Keep calling redirect_to_specific_clone even if
	TARGET_HAS_FMV_TARGET_ATTRIBUTE is defined.