RISC-V: Split parse_features_for_version with cl_target_option returned
In some new cases, we need to parse the features for a specific version
of the target and get the cl_target_option. This commit splits the
parse_features_for_version function to a new function that returns
the cl_target_option. The original function is kept for backward
compatibility.

Signed-off-by: Yangyu Chen <cyy@cyyself.name>

gcc/ChangeLog:

	* config/riscv/riscv.cc (parse_features_for_version): Split
	parse_features_for_version with cl_target_option returned.