Fortran: Do not make_decl_rtl in trans_function_start
This patch is a preparation for the function multi-versioning support
in Fortran. The function multi-versioning support requires changing the
change_decl_assembler_name to add the target suffix, which is not
allowed after make_decl_rtl is called, since the assembler name will be
in the constant pool.

This patch removes the make_decl_rtl call in trans_function_start, and
it will finally be called in all my testcases since we will check it
whenever using the DECL_RTL macro.

Signed-off-by: Yangyu Chen <cyy@cyyself.name>

gcc/fortran/ChangeLog:

	* trans-decl.cc (trans_function_start): Remove call make_decl_rtl.